Apply __attribute__((packed)) to TestStruct
These changes were tested on Arduino Uno with and without the __attribute__, and using
SoftwareSerial for the rx_datum and rx_data tests, and Serial for the tx_data/datum tests,
to ensure they didn't change behavior. They were tested on ESP32 Feather to check that the
test failed without the __attribute__ and passed with it. HardwareSeral(1) was used instead
of Serial1. Then the esp32-specific (and for Uno, the SoftwareSerial-specific) changes were
backed out, leaving only the __attribute__ change.

These changes address issue #51 in pySerialTransfer, by fixing TestStruct in SerialTransfer.
Another pull request will make the corresponding changes to examples in pySerialTransfer.